Upgraded build_pack_utils

+++ lib/build_pack_utils/cloudfoundry.py
@@ -37,7 +37,8 @@ class CloudFoundryUtil(object):
 
     @staticmethod
     def load_json_config_file_from(folder, cfgFile):
-        return CloudFoundryUtil.load_json_config_file(os.path.join(folder, cfgFile))
+        return CloudFoundryUtil.load_json_config_file(os.path.join(folder,
+                                                                   cfgFile))
 
     @staticmethod
     def load_json_config_file(cfgPath):
@@ -124,11 +125,12 @@ class CloudFoundryInstaller(object):
                           uses the bpFile path.
         """
         fullPathFrom = os.path.join(self._ctx['BP_DIR'], bpFile)
-        fullPathTo = os.path.join(
-            self._ctx['BUILD_DIR'],
-            ((toLocation is None) and bpFile or toLocation))
-        self._safe_makedirs(os.path.dirname(fullPathTo))
-        shutil.copy(fullPathFrom, fullPathTo)
+        if os.path.exists(fullPathFrom) and os.path.isfile(fullPathFrom):
+            fullPathTo = os.path.join(
+                self._ctx['BUILD_DIR'],
+                ((toLocation is None) and bpFile or toLocation))
+            self._safe_makedirs(os.path.dirname(fullPathTo))
+            shutil.copy(fullPathFrom, fullPathTo)
 
     def install_from_application(self, cfgFile, toLocation):
         """Copy file from one place to another in the application
@@ -141,6 +143,7 @@ class CloudFoundryInstaller(object):
                           relative to app droplet.
         """
         fullPathFrom = os.path.join(self._ctx['BUILD_DIR'], cfgFile)
-        fullPathTo = os.path.join(self._ctx['BUILD_DIR'], toLocation)
-        self._safe_makedirs(os.path.dirname(fullPathTo))
-        shutil.copy(fullPathFrom, fullPathTo)
+        if os.path.exists(fullPathFrom) and os.path.isfile(fullPathFrom):
+            fullPathTo = os.path.join(self._ctx['BUILD_DIR'], toLocation)
+            self._safe_makedirs(os.path.dirname(fullPathTo))
+            shutil.copy(fullPathFrom, fullPathTo)
